replaced all old dynamo db utils with new ones
