--- backend/aws_helpers/dynamo_db_utils/__init__.py
@@ -1,1 +1,0 @@
-__version__ = "0.1"
