+++ backend/aws_helpers/dynamo_db_utils/execution_db.py
@@ -1,221 +1,85 @@
-import json
-
-from dataclasses import dataclass
-from datetime import datetime
-
-from backend.aws_helpers.dynamo_db_utils.base_db import (
-    BaseData,
-    BaseDDBUtil,
-    enumclass,
-    changevar,
+from backend.aws_helpers.dynamo_db_utils.dynamo_db_utils import (
+    create_dynamo_item,
+    get_dynamo_item_by_id,
+    get_dynamo_items_by_gsi,
+    update_dynamo_item,
 )
-from backend.common.constants import EXECUTION_TABLE_NAME, AWS_REGION
-from boto3.dynamodb.conditions import Key
-from backend.common.utils import get_current_timestamp
-from typing import Union
+import random
+from datetime import datetime
+from dataclasses import dataclass
+
+TABLE_NAME = "execution-table"
 
 
 @dataclass
-class ExecutionData(BaseData):
+class ExecutionData:
     """Data class to hold the attribute values of a record of the execution-table DynamoDB table"""
 
     execution_id: str
-    user_id: str = None
-    name: str = None
-    timestamp: str = None
-    data_source: str = None
-    status: str = None
-    progress: int = None
+    data_source: str
+    name: str
+    status: str
+    timestamp: str
+    user_id: str
+    progress: int = 0
+    results: dict = None
+    metadata: dict = None
 
 
-@enumclass(
-    DataClass=ExecutionData,
-    data_source=[
-        "TABULAR",
-        "PRETRAINED",
-        "IMAGE",
-        "AUDIO",
-        "TEXTUAL",
-        "CLASSICAL_ML",
-        "OBJECT_DETECTION",
-    ],
-    status=["QUEUED", "STARTING", "UPLOADING", "TRAINING", "SUCCESS", "ERROR"],
-)
-class ExecutionEnums:
-    """Class that holds the enums associated with the ExecutionDDBUtil class. It includes:
-    ExecutionEnums.Attribute - Enum that defines the schema of the execution-table. It holds the attribute names of the table
-    ExecutionEnums.Execution_Source - Enum that defines the categorical values associated with the 'execution_source' attribute
-    """
-
-    pass
-
-
-@changevar(
-    DataClass=ExecutionData, EnumClass=ExecutionEnums, partition_key="execution_id"
-)
-class ExecutionDDBUtil(BaseDDBUtil):
-    """Class that interacts with AWS DynamoDB to manipulate information stored in the execution-table DynamoDB table"""
-
-    pass
-
-
-def getUserExecutionsData(execution_id: str) -> str:
+def getExecutionData(execution_id: str) -> dict:
     """
     Retrieves an entry from the `execution-table` DynamoDB table given an `execution_id`. Example output: {"execution_id": "blah", "user_id": "blah", "name": "blah", "timestamp": "blah", "data_source": "TABULAR", "status": "QUEUED", "progress": 1}
 
     @param execution_id: The execution_id of the entry to be retrieved
-    @return: A JSON string of the entry retrieved from the table
+    @return: A JSON dict of the entry retrieved from the table
     """
-    dynamoTable = ExecutionDDBUtil(EXECUTION_TABLE_NAME, AWS_REGION)
-    record = dynamoTable.get_record(execution_id)
-    return json.dumps(record.__dict__)
+    record = get_dynamo_item_by_id(TABLE_NAME, execution_id)
+    return record
 
 
-def createUserExecutionsData(entryData: dict) -> str:
-    """
-    Create a new entry corresponding to the given user_id.
-
-    @param **kwargs: execution_id and other table attributes to be created to the new entry e.g. user_id, if does not exist
-    @return: A JSON string of the entry retrieved or created from the table
-    """
-    required_keys = ["execution_id", "user_id"]
-    if not validate_keys(entryData, required_keys):
-        raise ValueError(f"Missing keys {required_keys} in request body")
-
-    dynamoTable = ExecutionDDBUtil(EXECUTION_TABLE_NAME, AWS_REGION)
-    newRecord = ExecutionData(**entryData)
-    dynamoTable.create_record(newRecord)
-    return json.dumps(newRecord.__dict__)
-
-
-def updateUserExecutionsData(requestData: dict) -> str:
+def updateExecutionData(execution_id: str, requestData: dict) -> bool:
     """
     Updates an entry from the `execution-table` DynamoDB table given an `execution_id`.
-    @param requestData: A dictionary containing the execution_id and other table attributes to be updated, with user_id as a required field
-    @return a success status message if the update is successful
-    """
-
-    required_keys = ["execution_id"]
-    if not validate_keys(requestData, required_keys):
-        raise ValueError(f"Missing keys {required_keys} in request body")
-
-    dynamoTable = ExecutionDDBUtil(EXECUTION_TABLE_NAME, AWS_REGION)
-    execution_id = requestData["execution_id"]
-    updatedRecord = ExecutionData(**requestData).__dict__
-    updatedRecord.pop("execution_id")
-    dynamoTable.update_record(execution_id, **updatedRecord)
-    return '{"status": "success"}'
-
-
-def getAllUserExecutionsData(user_id: str) -> str:
-    """
-    Retrieves an entry from the `execution-table` DynamoDB table given an `execution_id`. Example output: {"execution_id": "blah", "user_id": "blah", "name": "blah", "timestamp": "blah", "data_source": "TABULAR", "status": "QUEUED", "progress": 1}
-
-    @param execution_id: The execution_id of the entry to be retrieved
-    @return: A JSON string of the entry retrieved from the table
-    """
-    dynamoTable = ExecutionDDBUtil(EXECUTION_TABLE_NAME, AWS_REGION)
-    response = dynamoTable.table.query(
-        IndexName="user_id", KeyConditionExpression=Key("user_id").eq(user_id)
-    )
-    items = response["Items"]
-    record = []
-    for item in items:
-        data_source, name, progress, status, timestamp, execution_id = (
-            item["data_source"],
-            item["name"],
-            item["progress"],
-            item["status"],
-            item["timestamp"],
-            item["execution_id"],
-        )
-        record.append(
-            {
-                "data_source": data_source,
-                "name": name,
-                "progress": float(progress),
-                "status": status,
-                "timestamp": timestamp,
-                "execution_id": execution_id,
-            }
-        )
-
-    while "LastEvaluatedKey" in response:
-        key = response["LastEvaluatedKey"]
-        response = dynamoTable.table.query(
-            KeyConditionExpression=Key("user_id").eq(user_id), ExclusiveStartKey=key
-        )
-        for item in items:
-            data_source, name, progress, status, timestamp, execution_id = (
-                item["data_source"],
-                item["name"],
-                item["progress"],
-                item["status"],
-                item["timestamp"],
-                item["execution_id"],
-            )
-            record.append(
-                {
-                    "data_source": data_source,
-                    "name": name,
-                    "progress": float(progress),
-                    "status": status,
-                    "timestamp": timestamp,
-                    "execution_id": execution_id,
-                }
-            )
-    return json.dumps(record)
-
-
-def updateUserExecutionsData_(requestData: dict) -> str:
-    """
-    Updates an entry from the `execution-table` DynamoDB table given an `execution_id`.
-
-    @param requestData: A dictionary containing the execution_id and other table attributes to be updated, with user_id as a required field
-    @return a success status message if the update is successful
-    """
-
-    required_keys = ["execution_id"]
-    if not validate_keys(requestData, required_keys):
-        raise ValueError(f"Missing keys {required_keys} in request body")
-
-    dynamoTable = ExecutionDDBUtil(EXECUTION_TABLE_NAME, AWS_REGION)
-    execution_id = requestData["execution_id"]
-    updatedRecord = ExecutionData(**requestData).__dict__
-    updatedRecord.pop("execution_id")
-    dynamoTable.update_record(execution_id, **updatedRecord)
-    return '{"status": "success"}'
-
-
-def updateStatus(execution_id: str, status: str, entryData: dict = None) -> str:
-    """
-    Updates the status of an entry from the `execution-table` DynamoDB table given an `execution_id`.
 
     @param execution_id: The execution_id of the entry to be updated
-    @param status: The new status of the entry
+    @param requestData: A dictionary containing the table attributes to be updated, not including execution_id
     @return a success status message if the update is successful
     """
-    try:
-        dynamoTable = ExecutionDDBUtil(EXECUTION_TABLE_NAME, AWS_REGION)
-        dynamoTable.update_record(
-            execution_id, status=status, timestamp=get_current_timestamp()
-        )
-        return '{"status": "success"}'
-    except:
-        entryData["status"] = status
-        createUserExecutionsData(entryData)
-        return '{"status": "success"}'
+    return update_dynamo_item(TABLE_NAME, execution_id, requestData)
 
 
-def validate_keys(requestData: dict, required_keys: list[str]) -> bool:
+def getAllUserExecutionData(user_id: str) -> list[dict]:
     """
-    Validates all required_keys in the requestData dictionary are present.
+    Retrieves all entries of this user from the `execution-table` DynamoDB table given an `execution_id`. Example output: [{"execution_id": "blah", "user_id": "blah", "name": "blah", "timestamp": "blah", "data_source": "TABULAR", "status": "QUEUED", "progress": 1}]
 
-    @param requestData: The dictionary to be checked
-    @param required_keys: The list of keys to be checked
-    @return a boolean value indicating whether all required_keys are present in the requestData dictionary
+    @param user_id: The user_id of the entries to be retrieved
+    @return: A list of all matching entries retrieved from the table
     """
-    for key in required_keys:
-        if key not in requestData.keys():
-            return False
-    return True
+    response = get_dynamo_items_by_gsi(TABLE_NAME, user_id)
+    return response
+
+
+def updateStatus(execution_id: str, status: str, entryData: dict = None) -> bool:
+    """
+    Updates the status of a trainspace entry in the `trainspace` DynamoDB table given a `execution_id`. Also updates the entry with the given entryData if provided.
+
+    @param execution_id: The execution_id of the entry to be updated
+    @param status: The status to be updated
+    @param entryData: The entry to be updated (if any)
+    @return True if the update is successful
+    """
+    if entryData is None:
+        entryData = {}
+    entryData["status"] = status
+    return updateExecutionData(execution_id, entryData)
+
+
+def createExecutionData(execution_data: ExecutionData) -> bool:
+    """
+    Create a new entry corresponding to the given user_id. Replaces any existing entry with the same execution_id.
+
+    @param execution_data: execution_id and other table attributes to be created to the new entry e.g. user_id, if does not exist
+    @return: True if the creation is successful
+    """
+
+    return create_dynamo_item(TABLE_NAME, execution_data.__dict__)
