+++ backend/aws_helpers/dynamo_db_utils/trainspace_db.py
@@ -1,234 +1,85 @@
-import json
-
-from dataclasses import dataclass
-from datetime import datetime
-
-from backend.aws_helpers.dynamo_db_utils.base_db import (
-    BaseData,
-    BaseDDBUtil,
-    enumclass,
-    changevar,
+from backend.aws_helpers.dynamo_db_utils.dynamo_db_utils import (
+    create_dynamo_item,
+    get_dynamo_item_by_id,
+    get_dynamo_items_by_gsi,
+    update_dynamo_item,
 )
-from backend.common.constants import TRAINSPACE_TABLE_NAME, AWS_REGION
-from boto3.dynamodb.conditions import Key
-from backend.common.utils import get_current_timestamp
-from typing import Any, Union
+import random
+from datetime import datetime
+from dataclasses import dataclass
 
-PRIMARY_KEY = "trainspace_id"
-
-REQUIRED_KEYS = [
-    "trainspace_id",
-    "uid",
-]
+TABLE_NAME = "trainspace"
 
 
 @dataclass
-class DatasetData(BaseData):
-    name: str
-    is_default_dataset: bool
-
-
-@dataclass
-class ReviewData(BaseData):
-    notification_email: Union[str, None]
-    notification_phone_number: Union[str, None]
-
-
-@dataclass
-class TrainspaceData(BaseData):
-    """Data class to hold the attribute values of a record of the execution-table DynamoDB table"""
+class TrainspaceData:
+    """Data class to hold the attribute values of a record of the trainspace DynamoDB table"""
 
     trainspace_id: str
     uid: str
-    name: str
-    data_source: str
-    # train_parameters: str = None
-    # train_results: str = None
+    created: str = None
+    data_source: str = None
+    dataset_data: dict = None
+    name: str = None
+    parameters_data: dict = None
+    review_data: str = None
+    status: str = None
 
 
-@dataclass
-class LayerData(BaseData):
-    value: str
-    parameters: list[Any]
-
-
-@dataclass
-class TabularParametersData(BaseData):
-    target_col: str
-    features: list[str]
-    problem_type: str
-    criterion: str
-    optimizer_name: str
-    shuffle: bool
-    epochs: int
-    test_size: float
-    batch_size: int
-    layers: list[LayerData]
-
-
-@dataclass
-class TabularData(TrainspaceData):
-    dataset_data: DatasetData
-    parameters_data: TabularParametersData
-    review_data: ReviewData
-
-
-@enumclass(
-    DataClass=TrainspaceData,
-    data_source=[
-        "TABULAR",
-        "PRETRAINED",
-        "IMAGE",
-        "AUDIO",
-        "TEXTUAL",
-        "CLASSICAL_ML",
-        "OBJECT_DETECTION",
-    ],
-    # status=["QUEUED", "STARTING", "UPLOADING", "TRAINING", "SUCCESS", "ERROR"],
-)
-class TrainspaceEnums:
-    """Class that holds the enums associated with the ExecutionDDBUtil class. It includes:
-    ExecutionEnums.Attribute - Enum that defines the schema of the execution-table. It holds the attribute names of the table
-    ExecutionEnums.Execution_Source - Enum that defines the categorical values associated with the 'execution_source' attribute
+def getTrainspaceData(trainspace_id: str) -> dict:
     """
+    Retrieves an entry from the `trainspace` DynamoDB table given an `trainspace_id`. Example output: {"trainspace_id": "blah", "user_id": "blah", "name": "blah", "timestamp": "blah", "data_source": "TABULAR", "status": "QUEUED", "progress": 1}
 
-    pass
-
-
-@changevar(
-    DataClass=TrainspaceData, EnumClass=TrainspaceEnums, partition_key=PRIMARY_KEY
-)
-class TrainspaceDDBUtil(BaseDDBUtil):
-    """Class that interacts with AWS DynamoDB to manipulate information stored in the execution-table DynamoDB table"""
-
-    pass
-
-
-def getTrainspaceData(trainspace_id: str) -> str:
-    """
-    Retrieves an entry from the `execution-table` DynamoDB table given an `execution_id`. Example output: {"execution_id": "blah", "user_id": "blah", "name": "blah", "timestamp": "blah", "data_source": "TABULAR", "status": "QUEUED", "progress": 1}
-
-    @param execution_id: The execution_id of the entry to be retrieved
+    @param trainspace_id: The trainspace_id of the entry to be retrieved
     @return: A JSON string of the entry retrieved from the table
     """
-    dynamoTable = TrainspaceDDBUtil(TRAINSPACE_TABLE_NAME, AWS_REGION)
-    record = dynamoTable.get_record(trainspace_id)
-    return json.dumps(record.__dict__)
+    record = get_dynamo_item_by_id(TABLE_NAME, trainspace_id)
+    return record
 
 
-def createTrainspaceData(trainspace_data: TrainspaceData) -> None:
+def updateTrainspaceData(trainspace_id: str, requestData: dict) -> bool:
     """
-    Create a new entry corresponding to the given user_id.
+    Updates an entry from the `trainspace` DynamoDB table given an `trainspace_id`.
 
-    @param **kwargs: execution_id and other table attributes to be created to the new entry e.g. user_id, if does not exist
-    @return: A JSON string of the entry retrieved or created from the table
+    @param trainspace_id: The trainspace_id of the entry to be updated
+    @param requestData: A dictionary containing the other table attributes to be updated, not including trainspace_id
+    @return True if the update is successful
+    """
+    return update_dynamo_item(TABLE_NAME, trainspace_id, requestData)
+
+
+def getAllUserTrainspaceData(user_id: str) -> list[dict]:
+    """
+    Retrieves all entries of this user from the `trainspace` DynamoDB table given an `trainspace_id`. Example output: [{"trainspace_id": "blah", "user_id": "blah", "name": "blah", "timestamp": "blah", "data_source": "TABULAR", "status": "QUEUED", "progress": 1}]
+
+    @param trainspace_id: The trainspace_id of the entry to be retrieved
+    @return: A list of all matching entries retrieved from the table
+    """
+    response = get_dynamo_items_by_gsi(TABLE_NAME, user_id)
+    return response
+
+
+def updateStatus(trainspace_id: str, status: str, entryData: dict = None) -> bool:
+    """
+    Updates the status of a trainspace entry in the `trainspace` DynamoDB table given a `trainspace_id`. Also updates the entry with the given entryData if provided.
+
+    @param trainspace_id: The trainspace_id of the entry to be updated
+    @param status: The status to be updated
+    @param entryData: The entry to be updated (if any)
+    @return True if the update is successful
+    """
+    if entryData is None:
+        entryData = {}
+    entryData["status"] = status
+    return updateTrainspaceData(trainspace_id, entryData)
+
+
+def createTrainspaceData(trainspace_data: TrainspaceData) -> bool:
+    """
+    Create a new entry or replaces an existing entry table according to the `trainspace_id`.
+
+    @param trainspace_data: trainspace_id and other table attributes to be created or updated if the entry already exists
+    @return: True if the creation or update is successful
     """
 
-    dynamoTable = TrainspaceDDBUtil(TRAINSPACE_TABLE_NAME, AWS_REGION)
-    dynamoTable.create_record(trainspace_data)
-    # return json.dumps(trainspace_data.__dict__)
-
-
-def updateTrainspaceData(requestData: dict) -> str:
-    """
-    Updates an entry from the `execution-table` DynamoDB table given an `execution_id`.
-    @param requestData: A dictionary containing the execution_id and other table attributes to be updated, with user_id as a required field
-    @return a success status message if the update is successful
-    """
-    if not validate_keys(requestData, REQUIRED_KEYS):
-        raise ValueError(f"Missing keys {REQUIRED_KEYS} in request body")
-
-    dynamoTable = TrainspaceDDBUtil(TRAINSPACE_TABLE_NAME, AWS_REGION)
-    trainspace_id = requestData[PRIMARY_KEY]
-    updatedRecord = TrainspaceData(**requestData).__dict__
-    updatedRecord.pop(PRIMARY_KEY)
-    dynamoTable.update_record(trainspace_id, **updatedRecord)
-    return '{"status": "success"}'
-
-
-def getAllUserTrainspaceData(user_id: str) -> str:
-    """
-    Retrieves an entry from the `execution-table` DynamoDB table given an `execution_id`. Example output: {"execution_id": "blah", "user_id": "blah", "name": "blah", "timestamp": "blah", "data_source": "TABULAR", "status": "QUEUED", "progress": 1}
-
-    @param execution_id: The execution_id of the entry to be retrieved
-    @return: A JSON string of the entry retrieved from the table
-    """
-    dynamoTable = TrainspaceDDBUtil(TRAINSPACE_TABLE_NAME, AWS_REGION)
-    response = dynamoTable.table.query(
-        IndexName="uid", KeyConditionExpression=Key("uid").eq(user_id)
-    )
-    items = response["Items"]
-    record = []
-    for item in items:
-        record.append(
-            {
-                "trainspace_id": item["trainspace_id"],
-                "name": item["name"],
-                "training_file": item["training_file"],
-                "uid": item["uid"],
-                "step": item["step"],
-                "status": item["status"],
-                "created": item["created"],
-                "modified": item["modified"],
-                "train_model": item["train_model"],
-                "train_parameters": item["train_parameters"],
-                "train_results": item["train_results"],
-            }
-        )
-
-    while "LastEvaluatedKey" in response:
-        key = response["LastEvaluatedKey"]
-        response = dynamoTable.table.query(
-            KeyConditionExpression=Key("uid").eq(user_id), ExclusiveStartKey=key
-        )
-        for item in items:
-            record.append(
-                {
-                    "trainspace_id": item["trainspace_id"],
-                    "name": item["name"],
-                    "training_file": item["training_file"],
-                    "uid": item["uid"],
-                    "step": item["step"],
-                    "status": item["status"],
-                    "created": item["created"],
-                    "modified": item["modified"],
-                    "train_model": item["train_model"],
-                    "train_parameters": item["train_parameters"],
-                    "train_results": item["train_results"],
-                }
-            )
-    return json.dumps(record)
-
-
-def updateStatus(trainspace_id: str, status: str, entryData: dict = None) -> str:
-    """
-    Updates the status of an entry from the `execution-table` DynamoDB table given an `execution_id`.
-
-    @param execution_id: The execution_id of the entry to be updated
-    @param status: The new status of the entry
-    @return a success status message if the update is successful
-    """
-    try:
-        dynamoTable = TrainspaceDDBUtil(TRAINSPACE_TABLE_NAME, AWS_REGION)
-        dynamoTable.update_record(
-            trainspace_id, status=status, modified=get_current_timestamp()
-        )
-        return '{"status": "success"}'
-    except:
-        entryData["status"] = status
-        createTrainspaceData(entryData)
-        return '{"status": "success"}'
-
-
-def validate_keys(requestData: dict, required_keys: list[str]) -> bool:
-    """
-    Validates all required_keys in the requestData dictionary are present.
-
-    @param requestData: The dictionary to be checked
-    @param required_keys: The list of keys to be checked
-    @return a boolean value indicating whether all required_keys are present in the requestData dictionary
-    """
-    for key in required_keys:
-        if key not in requestData.keys():
-            return False
-    return True
+    return create_dynamo_item(TABLE_NAME, trainspace_data.__dict__)
