+++ backend/endpoints/train_endpoints.py
@@ -7,21 +7,12 @@ import uuid
 from flask import Blueprint
 from flask import request
 
-from backend.aws_helpers.dynamo_db_utils.execution_db import (
-    createUserExecutionsData,
-    updateStatus,
-)
 from backend.aws_helpers.dynamo_db_utils.trainspace_db import (
-    DatasetData,
-    LayerData,
-    ReviewData,
-    TabularData,
-    TabularParametersData,
+    TrainspaceData,
+    createTrainspaceData,
 )
 from backend.common.ai_drive import dl_tabular_drive, dl_img_drive, ml_drive
 from backend.common.constants import (
-    AWS_REGION,
-    TRAINSPACE_TABLE_NAME,
     UNZIPPED_DIR_NAME,
 )
 from backend.common.utils import *
@@ -67,47 +58,26 @@ def tabular_run():
     try:
         request_data = json.loads(request.data)
         id = str(uuid.uuid4())
-        tabular_data = TabularData(
+        tabular_data = TrainspaceData(
             trainspace_id=id,
             uid=request_data["user"]["uid"],
-            name=request_data["name"],
+            created=request_data["created"],
             data_source="TABULAR",
-            dataset_data=DatasetData(**request_data["dataset_data"]),
-            parameters_data=TabularParametersData(
-                target_col=request_data["parameters_data"]["target_col"],
-                features=request_data["parameters_data"]["features"],
-                problem_type=request_data["parameters_data"]["problem_type"],
-                criterion=request_data["parameters_data"]["criterion"],
-                optimizer_name=request_data["parameters_data"]["optimizer_name"],
-                shuffle=request_data["parameters_data"]["shuffle"],
-                epochs=request_data["parameters_data"]["epochs"],
-                test_size=request_data["parameters_data"]["test_size"],
-                batch_size=request_data["parameters_data"]["batch_size"],
-                layers=[
-                    LayerData(**layer)
-                    for layer in request_data["parameters_data"]["layers"]
-                ],
-            ),
-            review_data=ReviewData(**request_data["review_data"]),
+            dataset_data=request_data["dataset_data"],
+            name=request_data["name"],
+            parameters_data=request_data["parameters_data"],
+            review_data=request_data["review_data"],
         )
 
-        def to_decimal(o):
-            if isinstance(o, dict):
-                for k, v in o.items():
-                    if isinstance(v, float):
-                        o[k] = Decimal(str(v))
-                    else:
-                        to_decimal(v)
-            elif isinstance(o, list):
-                for v in o:
-                    to_decimal(v)
+        try:
+            createTrainspaceData(tabular_data)
+            print(id)
+            return send_success({"message": "success", "trainspace_id": id})
+        except Exception:
+            print(traceback.format_exc())
+            return send_traceback_error()
 
-        item = asdict(tabular_data)
-        to_decimal(item)
-        db = boto3.resource("dynamodb", AWS_REGION)
-        table = db.Table(TRAINSPACE_TABLE_NAME)
-        table.put_item(Item=item)
-        # createTrainspaceData(tabular_data)
+
         """ train_loss_results = dl_tabular_drive(
             user_arch, fileURL, params, json_csv_data_str, customModelName
         )
@@ -115,8 +85,7 @@ def tabular_run():
         print(train_loss_results)
         updateStatus(execution_id, "SUCCESS") """
         # return send_train_results(train_loss_results)
-        print(id)
-        return send_success({"message": "success", "trainspace_id": id})
+
 
     except Exception:
         # updateStatus(execution_id, "ERROR")
